+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/spring/SpringBooter.java
@@ -14,9 +14,13 @@
  */
 package com.ericsson.bss.cassandra.ecchronos.application.spring;
 
+import com.ericsson.bss.cassandra.ecchronos.application.ECChronosInternals;
 import com.ericsson.bss.cassandra.ecchronos.rest.RepairManagementRESTImpl;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.context.annotation.Import;
 
@@ -24,8 +28,16 @@ import org.springframework.context.annotation.Import;
 @Import(value = {RepairManagementRESTImpl.class})
 public class SpringBooter extends SpringBootServletInitializer
 {
+    private static final Logger LOG = LoggerFactory.getLogger(SpringBooter.class);
+
     public static void main(String[] args)
     {
-        SpringApplication.run(SpringBooter.class, args);
+        try{
+            SpringApplication.run(SpringBooter.class, args);
+        }
+        catch(Exception e){
+            LOG.error("Failed to initialize", e);
+            System.exit(1);
+        }
     }
 }